refactor(api): migrate controller findPendingInvitations
