+++ api/tests/team/unit/application/organization-invitation/organization-invitation.controller.test.js
@@ -1,4 +1,5 @@
 import { MissingQueryParamError } from '../../../../../lib/application/http-errors.js';
+import { usecases as usecasesLib } from '../../../../../lib/domain/usecases/index.js';
 import { organizationInvitationController } from '../../../../../src/team/application/organization-invitations/organization-invitation.controller.js';
 import { OrganizationInvitation } from '../../../../../src/team/domain/models/OrganizationInvitation.js';
 import { usecases } from '../../../../../src/team/domain/usecases/index.js';
@@ -45,6 +46,51 @@ describe('Unit | Team | Application | Controller | organization-invitation', fun
 
       // then
       expect(errorCatched).to.be.instanceof(MissingQueryParamError);
+    });
+  });
+
+  describe('#findPendingInvitations', function () {
+    const userId = 1;
+    let organization;
+    const resolvedOrganizationInvitations = 'organization invitations';
+    const serializedOrganizationInvitations = 'serialized organization invitations';
+
+    let request;
+    let dependencies;
+
+    beforeEach(function () {
+      organization = domainBuilder.buildOrganization();
+      request = {
+        auth: { credentials: { userId } },
+        params: { id: organization.id },
+      };
+
+      sinon.stub(usecasesLib, 'findPendingOrganizationInvitations');
+
+      const organizationInvitationSerializerStub = {
+        serialize: sinon.stub(),
+      };
+
+      dependencies = {
+        organizationInvitationSerializer: organizationInvitationSerializerStub,
+      };
+    });
+
+    it('calls the usecase to find pending invitations with organizationId', async function () {
+      usecasesLib.findPendingOrganizationInvitations.resolves(resolvedOrganizationInvitations);
+      dependencies.organizationInvitationSerializer.serialize.resolves(serializedOrganizationInvitations);
+
+      // when
+      const response = await organizationInvitationController.findPendingInvitations(request, hFake, dependencies);
+
+      // then
+      expect(usecasesLib.findPendingOrganizationInvitations).to.have.been.calledWithExactly({
+        organizationId: organization.id,
+      });
+      expect(dependencies.organizationInvitationSerializer.serialize).to.have.been.calledWithExactly(
+        resolvedOrganizationInvitations,
+      );
+      expect(response).to.deep.equal(serializedOrganizationInvitations);
     });
   });
