+++ api/tests/unit/application/organizations/organization-controller_test.js
@@ -284,51 +284,6 @@ describe('Unit | Application | Organizations | organization-controller', functio
     });
   });
 
-  describe('#findPendingInvitations', function () {
-    const userId = 1;
-    let organization;
-    const resolvedOrganizationInvitations = 'organization invitations';
-    const serializedOrganizationInvitations = 'serialized organization invitations';
-
-    let request;
-    let dependencies;
-
-    beforeEach(function () {
-      organization = domainBuilder.buildOrganization();
-      request = {
-        auth: { credentials: { userId } },
-        params: { id: organization.id },
-      };
-
-      sinon.stub(usecases, 'findPendingOrganizationInvitations');
-
-      const organizationInvitationSerializerStub = {
-        serialize: sinon.stub(),
-      };
-
-      dependencies = {
-        organizationInvitationSerializer: organizationInvitationSerializerStub,
-      };
-    });
-
-    it('should call the usecase to find pending invitations with organizationId', async function () {
-      usecases.findPendingOrganizationInvitations.resolves(resolvedOrganizationInvitations);
-      dependencies.organizationInvitationSerializer.serialize.resolves(serializedOrganizationInvitations);
-
-      // when
-      const response = await organizationController.findPendingInvitations(request, hFake, dependencies);
-
-      // then
-      expect(usecases.findPendingOrganizationInvitations).to.have.been.calledWithExactly({
-        organizationId: organization.id,
-      });
-      expect(dependencies.organizationInvitationSerializer.serialize).to.have.been.calledWithExactly(
-        resolvedOrganizationInvitations,
-      );
-      expect(response).to.deep.equal(serializedOrganizationInvitations);
-    });
-  });
-
   describe('#getOrganizationMemberIdentities', function () {
     it('should return all members identities of the organization serialized', async function () {
       // given
